// web-ui/api/types.gen.ts
export type CurrentUser = {
    readonly id: number;
    /**
     * Required. 30 characters or fewer. Letters, numbers and @/./+/-/_ characters
     */
    username: string;
    /**
     * Email address
     */
    readonly email: string;
    /**
     * Full name, will be shown in some places instead of username
     */
    display_name?: string;
    /**
     * Staff status
     * Designates whether the user can log into this admin site.
     */
    readonly is_staff: boolean;
    /**
     * Active
     * Designates whether this user should be treated as active. Unselect this instead of deleting accounts.
     */
    readonly is_active: boolean;
    picture: Picture;
};
// ...
export type Me = {
    user: CurrentUser | null;
    site: SiteConfig;
    can_invite: boolean;
    have_sessions: boolean;
    sso_login_url: string;
    sso_logout_url: string;
    plugins: Array<PluginInfo>;
};
// ...
export type Picture = {
    readonly id: number;
    readonly resized_url: string;
    readonly resized_png_url: string;
    readonly thumbnail_url: string;
    readonly thumbnail_png_url: string;
    caption: string;
    /**
     * User that owns/uploaded this picture
     */
    readonly user_id: number | null;
    /**
     * Keg this picture was taken with, if any.
     */
    readonly keg_id: number | null;
    /**
     * Session this picture was taken with, if any.
     */
    readonly session_id: number | null;
};

export type PictureUploadRequest = {
    image: string;
    caption?: string;
};

export type PluginData = {
    /**
     * Plugin short name
     */
    plugin_name: string;
    key: string;
    value: unknown;
};

export type PluginInfo = {
    short_name: string;
    name: string;
};